data structures,greedy,*1100

+++ Codeforces Round #805 (Div. 3)/C. Train and Queries.cpp
@@ -65,7 +65,7 @@ int sumOddLengthSubarrays(vector<int>& arr) {
         int i = _ + 1;
         if(i % 2) {
             sum += 
-             (arr[n - i - 1] * (((i * (i + 1) / 2) - (i / 2)) + ((n + 1) / 2 - i) * i));
+             (arr[n - i - 1] * (ll)(((i * (i + 1) / 2) - (i / 2)) + ((n + 1) / 2 - i) * i));
             // debug(i, (((i * (i + 1) / 2) - (i / 2)) + ((n + 1) / 2 - i) * i));
              debug(sum);
         }
